+++ src/main/java/com/strangegrotto/wealthdraft/assetimpls/bankaccount/BankAccountAssetSnapshot.java
@@ -1,5 +1,7 @@
 package com.strangegrotto.wealthdraft.assetimpls.bankaccount;
 
+import com.fasterxml.jackson.annotation.JsonCreator;
+import com.fasterxml.jackson.annotation.JsonProperty;
 import com.fasterxml.jackson.databind.annotation.JsonDeserialize;
 import com.google.common.annotations.VisibleForTesting;
 import com.strangegrotto.wealthdraft.WealthdraftImmutableStyle;
@@ -10,46 +12,58 @@ import org.immutables.value.Value;
 
 import java.math.BigDecimal;
 
-@WealthdraftImmutableStyle
-@Value.Immutable
-@JsonDeserialize(as = ImmBankAccountAssetSnapshot.class)
-public interface BankAccountAssetSnapshot extends AssetSnapshot {
-    int MONTHS_IN_YEAR = 12;
+public class BankAccountAssetSnapshot implements AssetSnapshot {
+    private static final int MONTHS_IN_YEAR = 12;
+
+    private final BigDecimal balance;
+    private final BigDecimal annualInterestRate;
+
+    @JsonCreator
+    public BankAccountAssetSnapshot(
+            @JsonProperty("balance") BigDecimal balance,
+            @JsonProperty("annualInterestRate") BigDecimal annualInterestRate
+    ) {
+        this.balance = balance;
+        this.annualInterestRate = annualInterestRate;
+    }
 
     @VisibleForTesting
-    BigDecimal getBalance();
+    BigDecimal getBalance() {
+        return this.balance;
+    }
 
     @VisibleForTesting
-    BigDecimal getAnnualInterestRate();
+    BigDecimal getAnnualInterestRate() {
+        return this.annualInterestRate;
+    }
 
     @Override
-    @Value.Derived
-    default BigDecimal getValue() {
+    public BigDecimal getValue() {
         return getBalance();
     }
 
 
     @Override
-    default BankAccountAssetSnapshot projectOneMonth() {
+    public BankAccountAssetSnapshot projectOneMonth() {
         // BigDecimal doesn't allow fractional exponents, so we drop down to Double (the loss of
         //  precision will be okay here)
-        double balanceDouble = getBalance().doubleValue();
-        double annualInterestRateDouble = getAnnualInterestRate().doubleValue();
+        double balanceDouble = this.balance.doubleValue();
+        double annualInterestRateDouble = this.annualInterestRate.doubleValue();
         double exponent = 1.0 / (double) MONTHS_IN_YEAR;
         double monthlyMultiplier = Math.pow(1 + annualInterestRateDouble, exponent);
         double newBalanceDouble = balanceDouble * monthlyMultiplier;
         BigDecimal newBalance = BigDecimal.valueOf(newBalanceDouble);
-        return ImmBankAccountAssetSnapshot.of(newBalance, getAnnualInterestRate());
+        return new BankAccountAssetSnapshot(newBalance, this.annualInterestRate);
     }
 
     @Override
-    default ValOrGerr<AssetSnapshot> applyChange(AssetChange change) {
+    public ValOrGerr<AssetSnapshot> applyChange(AssetChange change) {
         BankAccountAssetChange castedChange = (BankAccountAssetChange)change;
 
         var balanceModificationOpt = castedChange.getBalance();
-        var newBalance = getBalance();
+        var newBalance = this.balance;
         if (balanceModificationOpt.isPresent()) {
-            ValOrGerr<BigDecimal> newBalanceOrErr = balanceModificationOpt.get().apply(getBalance());
+            ValOrGerr<BigDecimal> newBalanceOrErr = balanceModificationOpt.get().apply(this.balance);
             if (newBalanceOrErr.hasGerr()) {
                 return ValOrGerr.propGerr(
                         newBalanceOrErr.getGerr(),
@@ -59,9 +73,9 @@ public interface BankAccountAssetSnapshot extends AssetSnapshot {
         }
 
         var interestRateModificationOpt = castedChange.getAnnualInterestRate();
-        var newInterestRate = getAnnualInterestRate();
+        var newInterestRate = this.annualInterestRate;
         if (interestRateModificationOpt.isPresent()) {
-            ValOrGerr<BigDecimal> newInterestRateOrErr = interestRateModificationOpt.get().apply(getAnnualInterestRate());
+            ValOrGerr<BigDecimal> newInterestRateOrErr = interestRateModificationOpt.get().apply(this.annualInterestRate);
             if (newInterestRateOrErr.hasGerr()) {
                 return ValOrGerr.propGerr(
                         newInterestRateOrErr.getGerr(),