+++ src/main/java/com/strangegrotto/wealthdraft/assetimpls/stock/StockAsset.java
@@ -1,41 +1,40 @@
-package com.strangegrotto.wealthdraft.assetimpls.bankaccount;
+package com.strangegrotto.wealthdraft.assetimpls.stock;
 
 import com.fasterxml.jackson.annotation.JsonCreator;
 import com.fasterxml.jackson.annotation.JsonProperty;
-import com.fasterxml.jackson.databind.annotation.JsonDeserialize;
-import com.strangegrotto.wealthdraft.WealthdraftImmutableStyle;
 import com.strangegrotto.wealthdraft.assetimpls.AssetTypeTagValue;
+import com.strangegrotto.wealthdraft.assetimpls.bankaccount.BankAccountAssetChange;
+import com.strangegrotto.wealthdraft.assetimpls.bankaccount.BankAccountAssetSnapshot;
 import com.strangegrotto.wealthdraft.assets.definition.AbstractAsset;
 import com.strangegrotto.wealthdraft.assets.definition.Asset;
-import com.strangegrotto.wealthdraft.assets.temporal.AssetSnapshot;
-import com.strangegrotto.wealthdraft.assets.definition.IntrinsicAssetTag;
 import com.strangegrotto.wealthdraft.assets.temporal.AssetChange;
+import com.strangegrotto.wealthdraft.assets.temporal.AssetSnapshot;
 import org.immutables.value.Value;
 
 import java.util.HashMap;
 import java.util.Map;
 
-public final class BankAccountAsset extends AbstractAsset {
+public final class StockAsset extends AbstractAsset {
     @JsonCreator
-    public BankAccountAsset(
+    public StockAsset(
             @JsonProperty("name") String name,
-            @JsonProperty("tags") Map<String, String> customTags
+            @JsonProperty("tags") Map<String, String> tags
     ) {
-        super(name, customTags);
+        super(name, tags);
     }
 
     @Override
     public Class<? extends AssetChange> getChangeType() {
-        return BankAccountAssetChange.class;
+        return null;
     }
 
     @Override
     public Class<? extends AssetSnapshot> getSnapshotType() {
-        return BankAccountAssetSnapshot.class;
+        return StockAssetSnapshot.class;
     }
 
     @Override
     public AssetTypeTagValue getAssetTypeTagValue() {
-        return AssetTypeTagValue.BANK_ACCOUNT;
+        return null;
     }
 }