+++ src/test/java/com/strangegrotto/wealthdraft/assetimpls/stock/StockAssetSnapshotTest.java
@@ -1,0 +1,26 @@
+package com.strangegrotto.wealthdraft.assetimpls.stock;
+
+import com.strangegrotto.wealthdraft.assetimpls.bankaccount.BankAccountAssetSnapshot;
+import com.strangegrotto.wealthdraft.assetimpls.bankaccount.ImmBankAccountAssetChange;
+import com.strangegrotto.wealthdraft.assetimpls.bankaccount.ImmBankAccountAssetSnapshot;
+import com.strangegrotto.wealthdraft.assets.temporal.AssetParameterChangeValueOperation;
+import com.strangegrotto.wealthdraft.assets.temporal.ImmAssetParameterChange;
+import org.junit.Assert;
+import org.junit.Test;
+
+import java.math.BigDecimal;
+
+// TODO Maybe make this an interface????
+public class StockAssetSnapshotTest {
+    @Test
+    public void testProjectionMath() {
+        // TODO
+        Assert.fail();
+    }
+
+    @Test
+    public void testChangeApplication() {
+        // TODO
+        Assert.fail();
+    }
+}