+++ src/test/java/com/strangegrotto/wealthdraft/assets/definition/AssetDefinitionsTest.java
@@ -2,7 +2,7 @@ package com.strangegrotto.wealthdraft.assets.definition;
 
 import com.fasterxml.jackson.databind.exc.ValueInstantiationException;
 import com.strangegrotto.wealthdraft.Main;
-import com.strangegrotto.wealthdraft.networth.history.AssetsHistoryFiles;
+import com.strangegrotto.wealthdraft.assetimpls.AssetTypeTagValue;
 import org.junit.Assert;
 import org.junit.Test;
 
@@ -11,7 +11,7 @@ import java.io.IOException;
 public class AssetDefinitionsTest {
     @Test
     public void testValidDeserialization() throws IOException {
-        AssetDefinitions definitions = parseAssetsFile(AssetDefinitionsFiles.EXAMPLE);
+        AssetDefinitions definitions = parseAssetsFile(AssetDefinitionsTestFiles.EXAMPLE);
         Assert.assertEquals(
                 ExpectedExampleAssetDefinitions.EXPECTED_ASSETS,
                 definitions.getAssets()
@@ -22,22 +22,60 @@ public class AssetDefinitionsTest {
         );
     }
 
+
+    @Test
+    public void testDeserializngEveryAsset() throws IOException {
+        var definitions = parseAssetsFile(AssetDefinitionsTestFiles.EVERY_ASSET_TYPE);
+        var assetsMap = definitions.getAssets();
+        var assets = assetsMap.values();
+
+        var expectedDistinctAssets = AssetTypeTagValue.values().length;
+
+        // Verify that we do indeed have one of every class
+        var distinctAssetClasses = assets.stream()
+                .map(Object::getClass)
+                .distinct()
+                .count();
+        Assert.assertEquals(expectedDistinctAssets, distinctAssetClasses);
+
+        // Verify no assets are reusing snapshot classes
+        var distinctSnapshotClasses = assets.stream()
+                .map(Asset::getSnapshotType)
+                .distinct()
+                .count();
+        Assert.assertEquals(expectedDistinctAssets, distinctSnapshotClasses);
+
+        // Verify no assets are reusing change classes
+        var distinctChangeClasses = assets.stream()
+                .map(Asset::getChangeType)
+                .distinct()
+                .count();
+        Assert.assertEquals(expectedDistinctAssets, distinctChangeClasses);
+
+        // Verify each asset has a unique asset type tag value
+        var distinctAssetTypeTags = assets.stream()
+                .map(Asset::getAssetTypeTagValue)
+                .distinct()
+                .count();
+        Assert.assertEquals(expectedDistinctAssets, distinctAssetTypeTags);
+    }
+
     @Test(expected = ValueInstantiationException.class)
     public void testErrorOnInvalidTagValue() throws IOException {
-        parseAssetsFile(AssetDefinitionsFiles.DISALLOWED_TAG_VALUES);
+        parseAssetsFile(AssetDefinitionsTestFiles.DISALLOWED_TAG_VALUES);
     }
 
     @Test(expected = ValueInstantiationException.class)
     public void testErrorOnCollisionWithIntrinsicTag() throws IOException {
-        parseAssetsFile(AssetDefinitionsFiles.INTRINSIC_TAG_COLLISION);
+        parseAssetsFile(AssetDefinitionsTestFiles.INTRINSIC_TAG_COLLISION);
     }
 
     @Test(expected = ValueInstantiationException.class)
     public void testErrorOnUnrecognizedTag() throws IOException {
-        parseAssetsFile(AssetDefinitionsFiles.UNRECOGNIZED_TAG);
+        parseAssetsFile(AssetDefinitionsTestFiles.UNRECOGNIZED_TAG);
     }
 
-    private static AssetDefinitions parseAssetsFile(AssetDefinitionsFiles testFile) throws IOException {
+    private static AssetDefinitions parseAssetsFile(AssetDefinitionsTestFiles testFile) throws IOException {
         var mapper = Main.getObjectMapper();
         var assetsUrl = testFile.getResource();
         return mapper.readValue(assetsUrl, AssetDefinitions.class);