+++ src/test/java/com/strangegrotto/wealthdraft/networth/history/AssetsHistoryTestFiles.java
@@ -3,9 +3,7 @@ package com.strangegrotto.wealthdraft.networth.history;
 import com.strangegrotto.wealthdraft.TestFileProvider;
 import com.strangegrotto.wealthdraft.TestResourceDirnames;
 
-import java.net.URL;
-
-public enum AssetsHistoryFiles implements TestFileProvider {
+public enum AssetsHistoryTestFiles implements TestFileProvider {
     NONEXISTENT_ASSET("nonexistent-asset.yml"),
     FUTURE_DATE("future-date.yml"),
     EXAMPLE(TestResourceDirnames.EXAMPLES.getDirname(), "assets-history.yml");
@@ -13,17 +11,23 @@ public enum AssetsHistoryFiles implements TestFileProvider {
     private final String containingDirname;
     private final String filename;
 
-    AssetsHistoryFiles(String containingDirname, String filename) {
+    AssetsHistoryTestFiles(String containingDirname, String filename) {
         this.containingDirname = containingDirname;
         this.filename = filename;
     }
 
-    AssetsHistoryFiles(String filename) {
+    AssetsHistoryTestFiles(String filename) {
         this(TestResourceDirnames.ASSETS_HISTORY_DESERIALIZATION_TESTS.getDirname(), filename);
     }
 
+
     @Override
-    public URL getResource() {
-        return getClass().getClassLoader().getResource(this.containingDirname + "/" + this.filename);
+    public String getContainingDirname() {
+        return this.containingDirname;
+    }
+
+    @Override
+    public String getFilename() {
+        return this.filename;
     }
 }